requirements.txt
# pip install -r requirements.txt

python-dotenv
openai
pyupbit
pyjwt
pandas
pandas_ta
schedule
streamlit
selenium